First version of background threading and discarding requests
Add a background thread for reading from stdin. This thread will write to a queue that the server will read on its main thread.

The server will look at the queue and determine if any of the messages can be discarded.

For example, a long sequence of textDocument/didChange requests do not need to be processed. Only the final one needs to be processed.

Similarly, other requests that may be in the queue before a textDocument/didChange may also be discarded.

Do a little work on only reparsing if needed.